refactor(FormGroup): client専用処理をclient/ErrorMessagesへ切り出しFormGroup自体から'use client'を除去する
ariaNotify呼び出しやMutationObserverなど、client専用APIに依存する
callbackRefのロジックとエラーメッセージ表示部分はErrorMessages側の
関心事であり、FormGroup本体には不要だった。

このロジックをclient/ErrorMessages.tsxへ移動することで、FormGroup
自体はuseMemoのみを使うモジュールになり'use client'を除去できる。
実際にビルド出力でFormGroup.jsに'use client'が付かなくなり、
client/ErrorMessages.jsにのみ付くことを確認済み。

hookが存在せずindex.tsからも公開されない内部コンポーネントのため、
client/直下にフラット配置する（CLAUDE.mdの判断基準）。
