+++ packages/smarthr-ui/src/components/FormGroup/client/index.ts
@@ -1,0 +1,1 @@
+export { ErrorMessages } from './ErrorMessages'
